Migrate NextStepClient to RestClient, unify type for counters, implement authentication tests, fix issues found by tests, clean up Next Step error handling in WebFlow

+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/entity/CredentialPolicyDetail.java
@@ -40,8 +40,8 @@ public class CredentialPolicyDetail {
     private Integer credentialLengthMin;
     private Integer credentialLengthMax;
     private String credentialAllowedChars;
-    private Long limitSoft;
-    private Long limitHard;
+    private Integer limitSoft;
+    private Integer limitHard;
     private int checkHistoryCount;
     private boolean rotationEnabled;
     private Integer rotationDays;
