+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/UpdateOtpPolicyResponse.java
@@ -33,7 +33,7 @@ public class UpdateOtpPolicyResponse {
     private String description;
     @NotNull
     private Integer length;
-    private Long attemptLimit;
+    private Integer attemptLimit;
     private Long expirationTime;
     private String genAlgorithm;
     @NotNull
